a64_audio_drv: switch to full packet scheduling
The original implementation used the 'packet half full' interrupt to
decide when to handle the next buffer. This, however, led to issues
under load where audible artifacts were produced and it took a non
deterministic time to recover from that situation.

Using the 'packet full' interrupt might also incur this problem but
is limited only to the time where the over-load is present. Together
with providing the driver enough CPU quota this allows for proper
playback even when the CPU utilization is high due to other components,
namely the 'cpu_burner'.

The commit contains also a check that should make the driver more robust
in the occurrence of spurious interrupts where the DMA buffer could be
updated prematurely.