preset/linphone: provide VoIP functionality
This preset bundles the 'linphonec' daemon and 'linphone-simple' GUI
to allow for convenient SIP telecommunication.

The configuration is stored in the '/recall/linphone' directory on
the used file-system (normally the SD card).

The GUI supports logging in via its 'Settings'-Page and these settings
are retained until a log-out is performed. Thus, on subsequent preset
usages, the log-in is done automatically.

At the top of the 'Main'-Page there are three favorite contacts that
can be populated freely. Pressing and holding the corresponding icon
will produce a pop-up dialog for removing the contact.

Out- and incoming calls are recorded in the contacts list. List items
may be removed by pressing and holding the corresponding item and
dragging it right (until a delete-icon is shown) and pressing the icon.
Contacts may also be added to the favorite list at the top
by dragging it left (until a add-icon is shown) and pressing the icon
if an unpopulated favorite contact is available.

Contacts are managed independently of the used SIP account.